fix: #493 undefined object issue
It seems that the provided object can be undefined.
This is a little bit strange to be honest, as the
api says the callback is non-null, but maybe I
do not understand it properly. We can safely
do the check. If _bshbCreated is not set the
further checks will still pass.